+++ Editor/Drawers/SceneTestCaseBehaviourDrawer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f306e5041d4713849838e9707ab07c5e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
