+++ Editor/Interfaces/ISceneTestCaseCollecter.cs
@@ -1,3 +1,5 @@
+// Copyright Edanoue, Inc. MIT License - see LICENSE.md
+
 #nullable enable
 
 using System.Collections.Generic;
